refactor: improve error handling and utils
